+++ R/tidyverse.R
@@ -21,7 +21,7 @@
 #' * `use_tidy_contributing()`: creates tidyverse contributing guidelines from
 #'    tidy_contributing template.
 #'
-#' * `use_issues()`: creates ISSUE_TEMPLATE.md from ISSUES.md template.
+#' * `use_tidy_issues()`: creates ISSUE_TEMPLATE.md from ISSUES.md template.
 #'
 #' @md
 #' @name tidyverse
@@ -139,7 +139,9 @@ use_tidy_contributing <- function() {
 #' @export
 #' @rdname tidyverse
 #' @inheritParams use_template
-use_issues <- function() {
+use_tidy_issues <- function() {
+  check_uses_github()
+
   use_template(
     "ISSUES.md",
     "ISSUE_TEMPLATE.md"